Redux Reducerの基盤作成
