+++ package.json
@@ -16,6 +16,8 @@
         "react-dom": "^17.0.2",
         "react-redux": "^7.2.0",
         "react-scripts": "4.0.3",
+        "redux": "^4.1.1",
+        "redux-thunk": "^2.3.0",
         "typescript": "~4.1.5"
     },
     "scripts": {
