+++ src/index.tsx
@@ -1,10 +1,15 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
-import App from './App'
-import { store } from './app/store'
+// import UseEffect from './UseEffect'
+// import { store } from './app/store'
 import { Provider } from 'react-redux'
-import * as serviceWorker from './serviceWorker'
+import App from './App'
+import { createStore } from 'redux'
+import rootReducer from './redux/reducers'
+
+const store = createStore(rootReducer)
+console.log(store.getState())
 
 ReactDOM.render(
     <React.StrictMode>
@@ -14,8 +19,3 @@ ReactDOM.render(
     </React.StrictMode>,
     document.getElementById('root')
 )
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()
