+++ src/redux/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import imageUrls from './imageUrls'
+
+export default combineReducers({
+    imageUrls,
+})
